+++ Changes.md
@@ -1,3 +1,4 @@
+* Display run name on unverified runs page
 * #1317 - handle no scores when creating top performance score reports
 * #1309 - Make edit note a panel on the right side
 * Add border under comments on the judge score summary display